QA INVOICE CALCULATION INTEGRATED 1

+++ apollo-api/src/main/kotlin/org/kebs/app/kotlin/apollo/api/ports/provided/dao/QADaoServices.kt
@@ -3941,7 +3941,7 @@ class QADaoServices(
                         permit,
                         user,
                         manufactureTurnOver,
-                        productsManufacture.size,
+                        productsManufacture.size.toLong(),
                         plantDetail
                     )
                 }
